VUEX #1: - form-components moved to views - some manipulations with router and auth promises
