+++ frontend/player/src/store/actions/auth/index.js
@@ -3,44 +3,44 @@ import authService from '@/services/auth'
 import {LOGIN_SUCCESS, LOGOUT_SUCCESS, LOGIN_OUT_FAIL, SIGNUP_FAIL} from '@/store/mutations/mutation-types'
 
 export const login = function ({commit}, creds, redirect) {
-  return authService.login(creds)
+  authService.login(creds)
     .then((user) => {
       commit(LOGIN_SUCCESS, {
         email: user.email,
         balance: user.balance
       });
       if (redirect) {
-        router.push(redirect)
+        router.push({name: redirect})
       }
     })
     .catch((error) => commit(LOGIN_OUT_FAIL, error))
 };
 
 export const signup = function ({commit}, creds, redirect) {
-  return authService.signup(creds)
+  authService.signup(creds)
     .then((user) => {
       commit(LOGIN_SUCCESS, {
         email: user.email,
         balance: user.balance
       });
       if (redirect) {
-        router.push(redirect)
+        router.push({name: redirect})
       }
     })
     .catch((error) => commit(SIGNUP_FAIL, error))
 };
 
 export const logout = function ({commit}) {
-  return authService.logout()
+  authService.logout()
     .then(() => {
       commit(LOGOUT_SUCCESS);
-      router.push('SRG')
+      router.push({name: 'SRG'})
     })
     .catch((error) => commit(LOGIN_OUT_FAIL, error))
 };
 
 export const checkAuth = function ({commit}) {
-  return authService.checkAuth()
+  authService.checkAuth()
     .then((response) => response.json())
     .then((user) =>
       commit(LOGIN_SUCCESS, {
