+++ frontend/player/src/views/App.vue
