+++ frontend/player/src/views/Home.vue
