+++ frontend/player/src/views/Signup.vue
